+++ apps/npm/svelte-check-darwin-arm64/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @rsvelte/svelte-check-darwin-arm64
 
+## 0.2.11
+
+## 0.2.10
+
 ## 0.2.9
 
 ## 0.2.8