+++ apps/npm/svelte-check-linux-x64-gnu/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @rsvelte/svelte-check-linux-x64-gnu
 
+## 0.2.11
+
+## 0.2.10
+
 ## 0.2.9
 
 ## 0.2.8